+++ frontend/src/router/index.js
@@ -13,6 +13,7 @@ const Campaigns = () => import('../views/Campaigns.vue')
 const ComparatorView = () => import('../views/ComparatorView.vue')
 const ImpactSimulator = () => import('../views/ImpactSimulator.vue')
 const WarRoom = () => import('../views/WarRoom.vue')
+const BudgetPlanner = () => import('../views/BudgetPlanner.vue')
 
 const routes = [
   {
@@ -75,6 +76,11 @@ const routes = [
     path: '/impact',
     name: 'ImpactSimulator',
     component: ImpactSimulator
+  },
+  {
+    path: '/budget-planner',
+    name: 'BudgetPlanner',
+    component: BudgetPlanner
   },
   {
     path: '/war-room',